requirements.txt
Django==5.2.8
djangorestframework==3.15.2
numpy==2.2.0
gunicorn==23.0.0
whitenoise==6.8.1
psycopg2-binary==2.9.10
dj-database-url==2.1.0
